+++ src/uniprotkb/adapters/proteinHighlights.ts
@@ -3,6 +3,7 @@ import { LocationDescriptorObject } from 'history';
 import { getEntryPath } from '../../app/config/urls';
 
 import { pluralise } from '../../shared/utils/utils';
+import { getEntrySectionNameAndId } from '../utils/entrySection';
 
 import { Namespace } from '../../shared/types/namespaces';
 import { TabLocation } from '../types/entry';
@@ -11,7 +12,7 @@ import {
   EntryType,
   getEntryTypeFromString,
 } from '../../shared/components/entry/EntryTypeIcon';
-import EntrySection, { getEntrySectionNameAndId } from '../types/entrySection';
+import EntrySection from '../types/entrySection';
 
 enum highlightSection {
   domains = 'domain',
